chore: refactor structure of pipeline related form file
