+++ src/components/forms/index.ts
@@ -1,10 +1,7 @@
-import CreatePipelineForm from "./CreatePipelineForm";
-import type { StepNumberState } from "./CreatePipelineForm/CreatePipelineForm";
-
 import OnboardingForm from "./OnboardingForm";
 
 export * from "./connector";
 export * from "./model";
+export * from "./pipeline";
 
-export { CreatePipelineForm, OnboardingForm };
-export type { StepNumberState };
+export { OnboardingForm };
