+++ src/components/forms/pipeline/CreatePipelineForm/CreatePipelineForm.tsx
@@ -1,7 +1,7 @@
 import { Dispatch, FC, SetStateAction } from "react";
 
 import CreatePipelineProgress from "@/components/ui/CreatePipelineProgress";
-import { FormikMultiStep } from "../../formik";
+import { FormikMultiStep } from "../../../formik";
 import SetupDestinationStep from "./SetupDestinationStep/SetupDestinationStep";
 import SetupModelStep from "./SetupModelStep";
 import SetupPipelineDetailsStep from "./SetupPipelineDetailsStep";
