+++ src/components/forms/pipeline/CreatePipelineForm/index.ts
@@ -1,1 +1,2 @@
 export { default } from "./CreatePipelineForm";
+export * from "./CreatePipelineForm";
